restructure WorkOrder classes
