+++ backend/src/main/java/com/interview/workorder/controller/WorkOrderController.java
@@ -1,5 +1,8 @@
-package com.interview.workorder;
+package com.interview.workorder.controller;
 
+import com.interview.workorder.request.WorkOrderRequest;
+import com.interview.workorder.response.WorkOrderResponse;
+import com.interview.workorder.service.WorkOrderService;
 import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.http.HttpStatus;
