+++ backend/src/main/java/com/interview/workorder/dao/WorkOrderRepository.java
@@ -1,5 +1,6 @@
-package com.interview.workorder;
+package com.interview.workorder.dao;
 
+import com.interview.workorder.entity.WorkOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WorkOrderRepository extends JpaRepository<WorkOrder, Long> {
