+++ backend/src/main/java/com/interview/workorder/service/WorkOrderService.java
@@ -1,8 +1,14 @@
-package com.interview.workorder;
+package com.interview.workorder.service;
 
-import com.interview.common.ResourceNotFoundException;
+import com.interview.common.error.ResourceNotFoundException;
 import java.time.LocalDateTime;
 import java.util.List;
+
+import com.interview.workorder.mapping.WorkOrderMapping;
+import com.interview.workorder.request.WorkOrderRequest;
+import com.interview.workorder.response.WorkOrderResponse;
+import com.interview.workorder.dao.WorkOrderRepository;
+import com.interview.workorder.entity.WorkOrder;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
@@ -17,31 +23,31 @@ public class WorkOrderService {
 
     public WorkOrderResponse create(WorkOrderRequest request) {
         WorkOrder workOrder = new WorkOrder();
-        mapRequest(workOrder, request);
+        WorkOrderMapping.mapRequest(workOrder, request);
         LocalDateTime now = LocalDateTime.now();
         workOrder.setCreatedAt(now);
         workOrder.setUpdatedAt(now);
 
-        return mapResponse(repository.save(workOrder));
+        return WorkOrderMapping.mapResponse(repository.save(workOrder));
     }
 
     public List<WorkOrderResponse> list() {
         return repository.findAll(Sort.by(Sort.Direction.ASC, "id"))
                 .stream()
-                .map(this::mapResponse)
+                .map(WorkOrderMapping::mapResponse)
                 .toList();
     }
 
     public WorkOrderResponse getById(Long id) {
-        return mapResponse(findByIdOrThrow(id));
+        return WorkOrderMapping.mapResponse(findByIdOrThrow(id));
     }
 
     public WorkOrderResponse update(Long id, WorkOrderRequest request) {
         WorkOrder existing = findByIdOrThrow(id);
-        mapRequest(existing, request);
+        WorkOrderMapping.mapRequest(existing, request);
         existing.setUpdatedAt(LocalDateTime.now());
 
-        return mapResponse(repository.save(existing));
+        return WorkOrderMapping.mapResponse(repository.save(existing));
     }
 
     public void delete(Long id) {
@@ -52,24 +58,5 @@ public class WorkOrderService {
     private WorkOrder findByIdOrThrow(Long id) {
         return repository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Work order with id " + id + " was not found"));
-    }
-
-    private void mapRequest(WorkOrder workOrder, WorkOrderRequest request) {
-        workOrder.setCustomerName(request.customerName());
-        workOrder.setVin(request.vin());
-        workOrder.setIssueDescription(request.issueDescription());
-        workOrder.setStatus(request.status());
-    }
-
-    private WorkOrderResponse mapResponse(WorkOrder workOrder) {
-        return new WorkOrderResponse(
-                workOrder.getId(),
-                workOrder.getCustomerName(),
-                workOrder.getVin(),
-                workOrder.getIssueDescription(),
-                workOrder.getStatus(),
-                workOrder.getCreatedAt(),
-                workOrder.getUpdatedAt()
-        );
     }
 }
